requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
Pillow>=8.0.0
openai>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
aiohttp>=3.8.0
opencv-python==4.6.0.66
numpy==1.26.4
paddleocr==2.7.0.3
matplotlib==3.8.3
faiss-cpu==1.7.4
scikit-learn==1.4.1.post1
pandas==2.2.1
pytest>=6.0.0
pytest-asyncio>=0.14.0 